+++ src/components/src/Echarts/LineChart.js
@@ -4,6 +4,16 @@ import ReactEcharts from 'echarts-for-react';
 import Basic from './Basic';
 
 export default class LineChart extends Basic {
+  componentWillReceiveProps(nextProps) {
+    if (this.chart) {
+      const chartInstance = this.chart.getEchartsInstance();
+      chartInstance.clear();
+      if (nextProps) {
+        chartInstance.setOption(this.getOption(nextProps));
+      }
+    }
+  }
+
   getOption(props) {
     const { color, config, onTooltipFormat } = props;
     const option = {
@@ -103,11 +113,12 @@ export default class LineChart extends Basic {
       option.tooltip.formatter = params => onTooltipFormat(params);
     }
     if (config.dataZoom) {
+      const { start } = config.dataZoom;
       option.dataZoom = [
         {
           show: true,
           realtime: true,
-          start: config.dataZoom.start || 30,
+          start: (start === null || start === undefined) ? 30 : start,
           end: config.dataZoom.end || 100,
         },
       ];
@@ -119,6 +130,9 @@ export default class LineChart extends Basic {
   render() {
     return (
       <ReactEcharts
+        ref={(ref) => {
+          this.chart = ref;
+        }}
         option={this.getOption(this.props)}
         style={this.props.style || { height: 250, width: '100%' }}
         notMerge